+++ src/transactions/manage_asset/UpdateAssetOpFrame.cpp
@@ -143,6 +143,10 @@ bool UpdateAssetOpFrame::doApply(Application& app, StorageHelper& storageHelper,
     }
     else
     {
+        if (ledgerManager.shouldUse(LedgerVersion::FIX_AUTOREVIEW))
+        {
+            autoreview = false;
+        }
 
         if (!ensureUpdateRequestValid(request))
         {