+++ src/xdr/types.h
@@ -36,6 +36,7 @@ enum class LedgerVersion : std::int32_t {
   FIX_MAX_SUBJECT_SIZE = 23,
   FIX_MOVEMENT_REVIEW = 24,
   FIX_SIGNATURE_CHECK = 25,
+  FIX_AUTOREVIEW = 26,
 };
 } namespace xdr {
 template<> struct xdr_traits<::stellar::LedgerVersion>
@@ -97,6 +98,8 @@ template<> struct xdr_traits<::stellar::LedgerVersion>
       return "FIX_MOVEMENT_REVIEW";
     case ::stellar::LedgerVersion::FIX_SIGNATURE_CHECK:
       return "FIX_SIGNATURE_CHECK";
+    case ::stellar::LedgerVersion::FIX_AUTOREVIEW:
+      return "FIX_AUTOREVIEW";
     default:
       return nullptr;
     }
@@ -128,7 +131,8 @@ template<> struct xdr_traits<::stellar::LedgerVersion>
       (int32_t)::stellar::LedgerVersion::MARK_ASSET_AS_DELETED,
       (int32_t)::stellar::LedgerVersion::FIX_MAX_SUBJECT_SIZE,
       (int32_t)::stellar::LedgerVersion::FIX_MOVEMENT_REVIEW,
-      (int32_t)::stellar::LedgerVersion::FIX_SIGNATURE_CHECK
+      (int32_t)::stellar::LedgerVersion::FIX_SIGNATURE_CHECK,
+      (int32_t)::stellar::LedgerVersion::FIX_AUTOREVIEW
     };
     return _xdr_enum_vec;
   }